telegram_bot/requirements.txt
aiogram==3.16.0
asyncpg==0.30.0
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.10.11
supabase==2.8.1